WIP adding MethodParameter to simplify method generation logic
- working on finding some failing ref objects
